+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+* Introduce new Tokenista::validate method and a validation result object to simplify
+  flows where the handling of an invalid token is the same but you want to e.g. log an
+  expired token differently to a tampered one.
+
 ## 1.2.0 (2019-04-03)
 
 * Drop support for php5 and test on php7